+++ db/services/userServices.js
@@ -1,25 +1,5 @@
 const User = require("../models/userModel");
 
-const createUser = async (body, password) => {
-  const user = await User.create(body, password);
-  return user;
-};
-
-const findUserByEmail = async (email) => {
-  const user = await User.findOne(email);
-  return user;
-};
-
-const userCollection = (body, avatar) => {
-  const user = new User(body, avatar);
-  return user;
-};
-
-const updateUserById = async (id, token) => {
-  const user = await User.findByIdAndUpdate(id, token);
-  return user;
-};
-
 const findUserById = async (id) => {
   const user = await User.findById(id);
   return user;
@@ -36,10 +16,6 @@ const createGoogleUser = async (email, name) => {
 };
 
 module.exports = {
-  createUser,
-  findUserByEmail,
-  userCollection,
-  updateUserById,
   findUserById,
   updateUserInfo,
   createGoogleUser,
