+++ middlewares/index.js
@@ -1,5 +1,6 @@
 const upload = require("./upload");
 const isValidId = require("./isValidId");
 const authenticate = require("./authenticate");
+const isEmptyBody = require("./isEmptyBody");
 
-module.exports = { upload, authenticate, isValidId };
+module.exports = { upload, authenticate, isValidId, isEmptyBody };
