+++ package.json
@@ -26,6 +26,7 @@
     "morgan": "1.10.0",
     "multer": "^1.4.5-lts.1",
     "multer-storage-cloudinary": "^4.0.0",
+    "nanoid": "^3.3.7",
     "nodemailer": "^6.9.8",
     "swagger-ui-express": "^5.0.0"
   },
